ELEC-427: Fix bugs, add serial

+++ pkg/msgs/can_msgs.go
@@ -6,12 +6,12 @@ import (
 
 // CAN describes a single message from the CAN bus.
 type CAN struct {
-	ID        uint16 `json:"id"`
+	ID        uint32 `json:"id"`
 	Timestamp uint64 `json:"timestamp"`
 	Data      uint64 `json:"data"`
 }
 
 // NewCAN creates a CAN struct, this should be used for all incoming messages.
-func NewCAN(id uint16, data uint64) CAN {
+func NewCAN(id uint32, data uint64) CAN {
 	return CAN{id, uint64(time.Now().UnixNano()) / uint64(time.Millisecond), data}
 }
